+++ packages/react-sdk/src/components/CallControls/ToggleVideoButton.tsx
@@ -55,17 +55,19 @@ export const ToggleVideoPublishingButton = (
     SfuModels.TrackType.VIDEO,
   );
 
-  const { toggleVideoMuteState: handleClick, isAwaitingApproval } =
+  const { toggleVideoMuteState: handleClick, isAwaitingPermission } =
     useToggleVideoMuteState();
 
   return (
     <Restricted requiredGrants={[OwnCapability.SEND_VIDEO]}>
       <PermissionNotification
         permission={OwnCapability.SEND_VIDEO}
-        isAwaitingApproval={isAwaitingApproval}
-        messageApproved="You can now share your video."
-        messageAwaitingApproval="Awaiting for an approval to share your video."
-        messageRevoked="You can no longer share your video."
+        isAwaitingApproval={isAwaitingPermission}
+        messageApproved={t('You can now share your video.')}
+        messageAwaitingApproval={t(
+          'Awaiting for an approval to share your video.',
+        )}
+        messageRevoked={t('You can no longer share your video.')}
       >
         <CompositeButton Menu={Menu} active={isVideoMute} caption={caption}>
           <IconButton