+++ packages/react-sdk/src/hooks/index.ts
@@ -2,3 +2,6 @@ export * from './useFloatingUIPreset';
 export * from './useScrollPosition';
 export * from './useToggleAudioMuteState';
 export * from './useToggleVideoMuteState';
+export * from './useToggleScreenShare';
+export * from './useToggleCallRecording';
+export * from './useRequestPermission';